Devel j stocker

+++ samples.py
@@ -9,41 +9,46 @@ import numpy as np
 import glob
 import line_comp as lc
 import gen_plots as gp
+
+from PIL import Image
+Image.MAX_IMAGE_PIXELS = None 
 #clear folders to generate new images
 
 
 def clear_folder(folder_path):
-    # Get all files in the folder
-    files = glob.glob(os.path.join(folder_path, "*"))
-    for f in files:
-        try:
-            os.remove(f)  # delete file
-        except Exception as e:
-            print(f"Could not delete {f}: {e}")
+    folder_path = os.path.abspath(folder_path)
+    base = os.path.abspath("./generated_images")
 
+    if folder_path == base:
+        raise RuntimeError("Refusing to clear generated_images root")
+
+    for f in glob.glob(os.path.join(folder_path, "*")):
+        if os.path.isfile(f):
+            os.remove(f)
 
 
 # Parameters
-std_dev_uni = 0.4
-AP_vol_50A = 0.4
-avg_rad_50A = np.linspace(50e-6, 100e-6, 200)
+std_dev_uni = 0.2
+AP_vol_50A = 0.5
+avg_rad_50A = np.linspace(40e-6, 140e-6, 200)
 
 avg_rad_50B = 70e-6  # fixed particle size
-AP_vol_50B = np.linspace(0.35, 0.60, 200)  # varying AP
+AP_vol_50B = np.linspace(0.35, 0.65, 200)  # varying AP
 
-std_dev_bi = [0.2, 0.5] #standard dev of coarse and fine groups
-mean_rad_bi_size = [np.linspace(90e-6, 110e-6, 50), [4e-6]*50]  # varying size, coarse/fine
-AP_fixed_bi = 0.5
+std_dev_bi = [0.2, 0.25] #standard dev of coarse and fine groups
+mean_rad_bi_size = [np.linspace(80e-6, 110e-6, 200), 25e-6]  # varying size, coarse/fine
+AP_fixed_bi = 0.55
 
-mean_rad_bi_AP = [110e-6, 4e-6]  # fixed size, coarse/fine
-AP_bi_var = np.linspace(0.5, 0.7, 50)  # varying AP
-mix_bi = 1/10 #1:10 coarse:fine
+mean_rad_bi_AP = [100e-6, 20e-6]  # fixed size, coarse/fine
+AP_bi_var = np.linspace(0.40, 0.55, 3)
+#AP_bi_var = np.linspace(0.35, 0.65, 5)  # varying AP
+mix_bi = 1/21 #coarse:fine
 
 img_size = 1
-physical_size = 30e-4
-max_attempts = 5000
+physical_size = 1e-2
+max_attempts = 100000
 
-interface_thickness = 5e-6 #5 um
+interface_thickness = 2e-6 #2 um
 
 folderA_png = "./generated_images/unimodal_varying_part_size/pngs"
 folderA_xyzr = "./generated_images/unimodal_varying_part_size/xyzrs"
@@ -58,43 +63,54 @@ folderC_xyzr = "./generated_images/bimodal_varying_part_size/xyzrs"
 folderD_png = "./generated_images/bimodal_varying_AP_ratio/pngs"
 folderD_xyzr = "./generated_images/bimodal_varying_AP_ratio/xyzrs"
 
+folderE_png = "./generated_images/test/pngs"
+folderE_xyzr = "./generated_images/test/xyzrs"
+
 folder_ignore = "./generated_images/ignore"
 
-resultsA = "./generated_images/unimodal_varying_part_size/results.txt"
-resultsB = "./generated_images/unimodal_varying_AP_ratio/results.txt"
-resultsC = "./generated_images/bimodal_varying_part_size/results.txt"
-resultsD = "./generated_images/bimodal_varying_AP_ratio/results.txt"
+resultsA = "./generated_images/unimodal_varying_part_size/results_uni_part_size.txt"
+resultsB = "./generated_images/unimodal_varying_AP_ratio/results_uni_AP_ratio.txt"
+resultsC = "./generated_images/bimodal_varying_part_size/results_bi_part_size.txt"
+resultsD = "./generated_images/bimodal_varying_AP_ratio/results_bi_AP_ratio.txt"
+resultsE = "./generated_images/test/results.txt"
 
 
 def reset():
-    clear_folder(folderA_png)
-    clear_folder(folderA_xyzr)
+    #clear_folder(folderA_png)
+    #clear_folder(folderA_xyzr)
 
-    clear_folder(folderB_png)
-    clear_folder(folderB_xyzr)
+    #clear_folder(folderB_png)
+    #clear_folder(folderB_xyzr)
 
-    clear_folder(folderC_png)
-    clear_folder(folderC_xyzr)
+    #clear_folder(folderC_png)
+    #clear_folder(folderC_xyzr)
 
-    clear_folder(folderD_png)
-    clear_folder(folderD_xyzr)
+    #clear_folder(folderD_png)
+    #clear_folder(folderD_xyzr)
+    
+    clear_folder(folderE_png)
+    clear_folder(folderE_xyzr)
 
-    os.makedirs(folderA_png, exist_ok=True)
-    os.makedirs(folderA_xyzr, exist_ok=True)
 
-    os.makedirs(folderB_png, exist_ok=True)
-    os.makedirs(folderB_xyzr, exist_ok=True)
+    #os.makedirs(folderA_png, exist_ok=True)
+    #os.makedirs(folderA_xyzr, exist_ok=True)
 
-    os.makedirs(folderC_png, exist_ok=True)
-    os.makedirs(folderC_xyzr, exist_ok=True)
+    #os.makedirs(folderB_png, exist_ok=True)
+    #os.makedirs(folderB_xyzr, exist_ok=True)
 
-    os.makedirs(folderD_png, exist_ok=True)
-    os.makedirs(folderD_xyzr, exist_ok=True)
+    #os.makedirs(folderC_png, exist_ok=True)
+    #os.makedirs(folderC_xyzr, exist_ok=True)
+
+    #os.makedirs(folderD_png, exist_ok=True)
+    #os.makedirs(folderD_xyzr, exist_ok=True)
+    
+    os.makedirs(folderE_png, exist_ok=True)
+    os.makedirs(folderE_xyzr, exist_ok=True)
 
     os.makedirs(folder_ignore, exist_ok=True)
 
 
-    for fp in [resultsA, resultsB, resultsC, resultsD]:
+    for fp in [resultsA, resultsB]:
         os.makedirs(os.path.dirname(fp), exist_ok=True)
 
 
@@ -112,14 +128,14 @@ def generateA():
         _, _, AP_achieved, _ = gm.gen_struct(
             temp_png, save_path_ignore, temp_xyzr,
             img_size, physical_size, radius, AP_vol_50A,
-            std_dev_uni, max_attempts, mode=1, max_tries=200
+            std_dev_uni, max_attempts, mode=1, max_tries=20
         )
 
         AP_str = str(int(AP_achieved * 10000))  # 4-digit AP fraction without leading 0
         radius_um = int(radius * 1e7) #3 digit radius
         final_png = os.path.join(folderA_png, f"uni_R{radius_um}um_AP{AP_str}.png")
         final_xyzr = os.path.join(folderA_xyzr, f"uni_R{radius_um}um_AP{AP_str}.xyzr")
-        ap, htpb, interface = lc.vert_avg(save_path_ignore)
+        ap, htpb, interface = lc.vert_avg_fast(save_path_ignore)
         results.append((radius_um, ap, htpb, interface))
         if os.path.exists(temp_png):
             os.replace(temp_png, final_png)
@@ -133,6 +149,7 @@ def generateA():
             f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}\n")
 
 
+
             
 def generateB():
     results = []
@@ -144,14 +161,14 @@ def generateB():
         _, _, AP_achieved, _ = gm.gen_struct(
             temp_png, save_path_ignore, temp_xyzr,
             img_size, physical_size, avg_rad_50B, AP_target,
-            std_dev_uni, max_attempts, mode=1, max_tries=200
+            std_dev_uni, max_attempts, mode=1, max_tries=5
         )
 
         AP_str = str(int(AP_achieved * 10000))
         radius_um = int(avg_rad_50B * 1e6)
         final_png = os.path.join(folderB_png, f"uni_AP{AP_str}_R{radius_um}um.png")
         final_xyzr = os.path.join(folderB_xyzr, f"uni_AP{AP_str}_R{radius_um}um.xyzr")
-        ap, htpb, interface = lc.vert_avg(save_path_ignore)
+        ap, htpb, interface = lc.vert_avg_fast(save_path_ignore)
         results.append((AP_achieved*100, ap, htpb, interface))
         if os.path.exists(temp_png):
             os.replace(temp_png, final_png)
@@ -164,28 +181,41 @@ def generateB():
         for item in results:
             f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}\n")
 
+
 def generateC():
     results = []
+    
     # --- C: Bimodal, varying particle size ---
     for i, coarse_radius in enumerate(mean_rad_bi_size[0]):
-        fine_radius = mean_rad_bi_size[1][i]  # pick corresponding fine particle
+        avg_radius = []
+        avg_radii = []
+        fine_radius = mean_rad_bi_size[1]  # pick corresponding fine particle
         radius_input = [float(coarse_radius), float(fine_radius)]  # pass as list for bimodal
 
         temp_png = os.path.join(folderC_png, f"temp_{i}.png")
         temp_xyzr = os.path.join(folderC_xyzr, f"temp_{i}.xyzr")
-
+        
         _, _, AP_achieved, _ = gm.gen_struct(
             temp_png, save_path_ignore, temp_xyzr,
             img_size, physical_size, radius_input, AP_fixed_bi,
-            std_dev_bi, max_attempts, mode=2, mix=mix_bi
+            std_dev_bi, max_attempts, mode=2, mix=mix_bi, max_tries=200
         )
+
+        #pull average radius from xyzr
+        with open(temp_xyzr, "r") as f:
+            for line in f:# 4th column = radius
+                parts = line.strip().split()
+                avg_radii.append(float(parts[3]))
+            image_avg_radius = np.mean(avg_radii)
+        avg_radius.append(image_avg_radius*1e6) #put in um
 
         AP_str = str(int(AP_achieved * 10000))  # 4-digit AP fraction
         radius_um = int(coarse_radius * 1e6) #3 digit radius
         final_png = os.path.join(folderC_png, f"bi_R{radius_um}um_AP{AP_str}.png")
         final_xyzr = os.path.join(folderC_xyzr, f"bi_R{radius_um}um_AP{AP_str}.xyzr")
-        ap, htpb, interface = lc.vert_avg(save_path_ignore)
-        results.append((AP_achieved, ap, htpb, interface))
+        ap, htpb, interface = lc.vert_avg_fast(save_path_ignore)
+        
+        results.append((AP_achieved, ap, htpb, interface, avg_radius[0]))
         if os.path.exists(temp_png):
             os.replace(temp_png, final_png)
         if os.path.exists(temp_xyzr):
@@ -193,33 +223,51 @@ def generateC():
         else:
             print(f"Warning: {temp_xyzr} was not created!")
     with open(resultsC, "w") as f:
-        f.write("AP_achieved, AP_vert, HTPB_vert, Interface_vert\n")  # optional header
+        f.write("AP_achieved, AP_vert, HTPB_vert, Interface_vert, Avg Radius\n")  # optional header
         for item in results:
-            f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}\n")
+            f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}, {item[4]}\n")
 
 def generateD():
     results = []
     # --- D: Bimodal, varying AP fraction ---
-    coarse_radius = mean_rad_bi_AP[0]
-    fine_radius = mean_rad_bi_AP[1]
-    radius_input = [coarse_radius, fine_radius]
+    r_target = 30e-6 #avg radius 70 micrometers  
+    mix = mix_bi                   # number fraction of coarse
+
+    coarse_radius = mean_rad_bi_AP[0]  # keep fixed
+    fine_radius = (r_target - mix * coarse_radius) / (1 - mix)
+
+    r_avg = mix_bi * coarse_radius + (1 - mix_bi) * fine_radius
+    r_avg_um = int(r_avg * 1e6)
+
 
     for i, AP_target in enumerate(AP_bi_var):
+        avg_radius = []
+        avg_radii = []
         temp_png = os.path.join(folderD_png, f"temp_{i}.png")
         temp_xyzr = os.path.join(folderD_xyzr, f"temp_{i}.xyzr")
+        radius_input = [coarse_radius, fine_radius]
 
+
+        
         _, _, AP_achieved, _ = gm.gen_struct(
             temp_png, save_path_ignore, temp_xyzr,
             img_size, physical_size, radius_input, AP_target,
-            std_dev_bi, max_attempts, mode=2, mix=mix_bi
+            std_dev_bi, max_attempts, mode=2, mix=mix_bi, max_tries=2
         )
+        
+        with open(temp_xyzr, "r") as f:
+            avg_radii = [float(line.strip().split()[3]) for line in f]
+        image_avg_radius = np.mean(avg_radii)
+        avg_radius.append(image_avg_radius*1e6)
+
 
         AP_str = str(int(AP_achieved * 10000))  # 4-digit AP fraction
-        radius_um = int(coarse_radius * 1e6)
-        final_png = os.path.join(folderD_png, f"bi_AP{AP_str}_R{radius_um}um.png")
-        final_xyzr = os.path.join(folderD_xyzr, f"bi_AP{AP_str}_R{radius_um}um.xyzr")
+        final_png = os.path.join(folderD_png, f"bi_AP{AP_str}_R{image_avg_radius*1e6:.1f}um.png")
+        final_xyzr = os.path.join(folderD_xyzr, f"bi_AP{AP_str}_R{image_avg_radius*1e6:.1f}um.xyzr")
+
+
         ap, htpb, interface = lc.vert_avg(save_path_ignore)
-        results.append((AP_achieved, ap, htpb, interface))
+        results.append((AP_achieved, ap, htpb, interface, avg_radius[0]))
         if os.path.exists(temp_png):
             os.replace(temp_png, final_png)
         if os.path.exists(temp_xyzr):
@@ -227,19 +275,69 @@ def generateD():
         else:
             print(f"Warning: {temp_xyzr} was not created!")
     with open(resultsD, "w") as f:
-        f.write("AP_achieved, AP_vert, HTPB_vert, Interface_vert\n")  # optional header
+        f.write("AP_achieved, AP_vert, HTPB_vert, Interface_vert, Avg Radius\n")  # optional header
         for item in results:
-            f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}\n")
+            f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}, {item[4]}\n")
 
 
 
+def tester_structures():
+    '''Generating random radii/%AP, put into text file, see if it matches with equation'''
+    #just going to test with unimodal to start
+    radius_m = np.random.uniform(40e-6, 140e-6, 2)
+    percent_ap = np.random.uniform(0.4, 0.6, 2)
+    results = []
+    MoE = []
+
+    for i, AP_target in enumerate(percent_ap):
+
+        R_m = radius_m[i]
+        R = R_m * 1e6
+
+        temp_png = os.path.join(folderE_png, f"temp_{i}.png")
+        temp_xyzr = os.path.join(folderE_xyzr, f"temp_{i}.xyzr")
+        
+        _, _, AP_achieved, _ = gm.gen_struct(
+            temp_png, save_path_ignore, temp_xyzr,
+            1, 1e-3, R_m, AP_target,
+            std_dev_uni, max_attempts=100000, mode=1, max_tries=1
+        )
+        
+
+        AP_str = str(int(AP_achieved * 10000))
+        radius_um = int(R_m * 1e6)
+        final_png = os.path.join(folderE_png, f"uni_AP{AP_str}_R{radius_um}um.png")
+        final_xyzr = os.path.join(folderE_xyzr, f"uni_AP{AP_str}_R{radius_um}um.xyzr")
+        ap, htpb, interface = lc.vert_avg_fast(save_path_ignore)
+        #I equation
+
+        I_theory = 1.652e-7 * R**4 - 8.324e-5 * R**3 \
+            + 1.598e-2 * R**2 - 1.379 * R \
+                + 27.61 * AP_achieved **2 - 3.989 * AP_achieved + 46.6469
+        
+        #equation's margin of error
+        MoE.append(np.abs(I_theory - interface)/interface * 100)
+        
+        results.append((AP_achieved*100, radius_um, interface, I_theory, MoE[i]))
+        if os.path.exists(temp_png):
+            os.replace(temp_png, final_png)
+        if os.path.exists(temp_xyzr):
+            os.replace(temp_xyzr, final_xyzr)
+        else:
+            print(f"Warning: {temp_xyzr} was not created!")
+        #put something with equation in here too...
+    with open(resultsE, "w") as f:
+        f.write("AP_achieved, Avg Radius (um), Interface Experimental, Interface from Eq., Eq. Margin of Error\n")  # optional header
+        for item in results:
+            f.write(f"{item[0]}, {item[1]}, {item[2]}, {item[3]}, {item[4]}\n")
+    print(np.mean(MoE))
     
+    
+
 
 
 if __name__ == "__main__":
     reset()
     #generateA()
-    generateB()
-    print('Success!')
-    #gp.plotA()
-    gp.plotB()
+    #generateB()
+    tester_structures()
